+++ CHANGELOG.md
@@ -4,15 +4,37 @@
 
 ### :wrench: Changed
 
-- Generate at least one non-NULL character in path parameters. [#2790](https://github.com/schemathesis/schemathesis/issues/2790)  
+- Generate at least one non-NULL character in path parameters. [#2790](https://github.com/schemathesis/schemathesis/issues/2790)
 - Improved failure messages for `unsupported_method` and `missing_required_header` checks.
+- Restore `Case.formatted_path`.
+- Update serialization error messages.
+- Update documentation links in error messages.
 
 ### :bug: Fixed
 
 - Internal error during negative testing in some cases on Open API 3.1.
 - **Pytest**: not using all built-in checks by default.
 - **Pytest**: Ignored `headers` from config file.
+- **Pytest**: Ignored `basic_auth` from config file.
 - Make header overrides case-insensitive.
+- Incorrect auth detection in `ignored_auth` when it is passed to `call` or `call_and_validate`. [#2846](https://github.com/schemathesis/schemathesis/issues/2846)
+- Incorrect check for cookie and query param based authorization in `ignored_auth`.
+
+### :fire: Removed
+
+- `@schema.override`. Use `parameters` configuration option instead.
+
+Change this:
+
+```python
+@schema.override(path_parameters={"user_id": 42})
+```
+
+To:
+
+```toml
+parameters = { user_id = 42 }
+```
 
 ## [4.0.0-alpha.12](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.11...v4.0.0-alpha.12) - 2025-06-02