+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.12...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.0-beta.1...HEAD) - TBD
+
+## [4.0.0-beta.1](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.12...v4.0.0-beta.1) - 2025-06-07
 
 ### :rocket: Added
 
@@ -30,10 +32,12 @@
 - Make header overrides case-insensitive.
 - Incorrect auth detection in `ignored_auth` when it is passed to `call` or `call_and_validate`. [#2846](https://github.com/schemathesis/schemathesis/issues/2846)
 - Incorrect check for cookie and query param based authorization in `ignored_auth`.
-- Do not include `parameters` in error message about missing HTTP method.
+- Do not include `parameters` in the error message about missing HTTP method.
 
 ### :fire: Removed
 
+- `Schema.add_link`. Adjust your API schema manually instead.
+- `Schema.configure`. Use the config file instead. If you used to pass `app` to `Schema.configure`, pass it to `Case.call` or `Case.call_and_validate` instead.
 - `@schema.override`. Use `parameters` configuration option instead.
 
 Change this: